API/requirements.txt
django
djangorestframework
environs
google-auth-oauthlib
google-generativeai
pdfminer.six
python-docx
faiss-cpu
python-dotenv
pdfplumber
vertexai
requests
djangorestframework-api-key
pandas